feat: made train dynamics parallelized
